Animate Orrery Project Start
Added the basic place holder for the planets and the sun. Added the scripts we learned in class in my project files. Started to use them on some of the planets to get things working somewhat but still need to put a lot of work into the code. Solid start for me personally but still a lot of work to do and learn.