Animated Orrery Project Update 1
Set it up now so that the planets spin and orbit at their own speeds. Also added some basic colors to the planets just as placeholders. Tried setting up the camera and making it able to follow planets when they are selected/clicked on but still trying to work out the kinks I am having with the camera currently. 

Made some more progress but still a lot of work to do.