Almost Final
Added orbit lines, added planet skins, added a skybox. 
You can also currently click on a planet and it will follow the orbit but have to find a way to get closer to the planet
Closer to Final

Added a time scale so you can reverse the orbit of the planets also added the moons. Getting closer to the final version but still not done.
Final 

System is still missing parts. I could not get the camera to move in on planet when you click on it so I added a scroll wheel script to still allow for the ability to zoom in on the planet. I also removed some assets that were taking up space on the file.